+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Add a stop button in the popup to abort a download (Linux/Mac only)
 - Move included JS to local app instead of accessing from a CDN
 - Make the simultaneous download limit apply to each unique domain
+- Support "portable" mode, reading gropple.yml from the current directory, if present
 
 ## [v0.5.3] - 2021-11-21